+++ lib/src/popup_container/marker_with_key.dart
@@ -13,9 +13,14 @@ class MarkerWithKey {
 
   static Marker unwrap(MarkerWithKey markerWithKey) => markerWithKey.marker;
 
+  // Uses the hashCode for equals comparison to allow Marker extensions to be
+  // considered equal to a normal Marker by overriding their hashCode to equal
+  // the hashCode of their container Marker. This is useful for FlutterMap
+  // plugins that need to wrap Markers with their own data but still have them
+  // match the user provided markers.
   @override
   bool operator ==(Object other) =>
-      other is MarkerWithKey && marker == other.marker;
+      other is MarkerWithKey && marker.hashCode == other.marker.hashCode;
 
   @override
   int get hashCode => marker.hashCode;